fix(runtime): preserve startup device precedence
